+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/FlutterOsmView.kt
@@ -210,6 +210,13 @@ class FlutterOsmView(
     private val mapListener by lazy {
         object : MapListener {
             override fun onScroll(event: ScrollEvent?): Boolean {
+                if (!isTracking && !isEnabled) {
+                    val hashMap = HashMap<String, Any?>()
+                    hashMap["bounding"] = map?.boundingBox?.toHashMap()
+                    hashMap["center"] = (map?.mapCenter as GeoPoint).toHashMap()
+                    methodChannel.invokeMethod("receiveRegionIsChanging", hashMap)
+                }
+
                 return true
             }
 
@@ -355,6 +362,9 @@ class FlutterOsmView(
                 "map#center" -> {
                     result.success((map?.mapCenter as GeoPoint).toHashMap())
                 }
+                "map#bounds" -> {
+                    getMapBounds(result = result)
+                }
                 "user#position" -> {
                     if (locationNewOverlay == null) {
                         locationNewOverlay = MyLocationNewOverlay(provider, map)
@@ -464,6 +474,11 @@ class FlutterOsmView(
             Log.e(e.cause.toString(), e.stackTraceToString())
             result.error("404", e.message, e.stackTraceToString())
         }
+    }
+
+    private fun getMapBounds(result: MethodChannel.Result) {
+        val bounds = map?.boundingBox ?: boundingWorldBox
+        result.success(bounds.toHashMap())
     }
 
 
@@ -1226,7 +1241,7 @@ class FlutterOsmView(
             roadManager = OSRMRoadManager(context, "json/application")
         roadManager?.let { manager ->
             manager.setMean(meanUrl)
-
+            var routePointsEncoded = ""
             job = scope?.launch(Default) {
                 val wayPoints = listPointsArgs.map {
                     GeoPoint(it["lat"]!!, it["lon"]!!)
@@ -1245,6 +1260,7 @@ class FlutterOsmView(
                 val road = manager.getRoad(roadPoints)
                 withContext(Main) {
                     if (road.mRouteHigh.size > 2) {
+                        routePointsEncoded = PolylineEncoder.encode(road.mRouteHigh, 10)
                         val polyLine = RoadManager.buildRoadOverlay(road)
                         createRoad(
                                 polyLine = polyLine,
@@ -1266,9 +1282,10 @@ class FlutterOsmView(
 
                         map!!.invalidate()
                     }
-                    result.success(HashMap<String, Double>().apply {
+                    result.success(HashMap<String, Any>().apply {
                         this["duration"] = road.mDuration
                         this["distance"] = road.mLength
+                        this["routePoints"] = routePointsEncoded
                     })
                 }