+++ flutter_osm_interface/lib/src/channel/osm_method_channel.dart
@@ -9,7 +9,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_osm_interface/src/types/geo_point.dart';
 import 'package:google_polyline_algorithm/google_polyline_algorithm.dart';
 import 'package:location/location.dart';
 import 'package:stream_transform/stream_transform.dart';
@@ -83,6 +82,11 @@ class MethodChannelOSM extends MobileOSMPlatform {
     return _events(idMap).whereType<UserLocationEvent>();
   }
 
+  @override
+  Stream<RegionIsChangingEvent> onRegionIsChangingListener(int idMap) {
+    return _events(idMap).whereType<RegionIsChangingEvent>();
+  }
+
   void setGeoPointHandler(int idMap) async {
     _channels[idMap]!.setMethodCallHandler((call) async {
       switch (call.method) {
@@ -110,6 +114,11 @@ class MethodChannelOSM extends MobileOSMPlatform {
           final result = call.arguments;
           _streamController
               .add(UserLocationEvent(idMap, GeoPoint.fromMap(result)));
+          break;
+        case "receiveRegionIsChanging":
+          final result = call.arguments;
+          _streamController
+              .add(RegionIsChangingEvent(idMap, Region.fromMap(result)));
           break;
       }
       return true;
@@ -631,6 +640,12 @@ class MethodChannelOSM extends MobileOSMPlatform {
   Future<GeoPoint> getMapCenter(int idMap) async {
     final result = await _channels[idMap]?.invokeMethod('map#center', []);
     return GeoPoint.fromMap(result);
+  }
+
+  @override
+  Future<BoundingBox> getBounds(int idOSM) async {
+    final Map mapBounds = await _channels[idOSM]?.invokeMethod('map#bounds');
+    return BoundingBox.fromMap(mapBounds);
   }
 }