+++ flutter_osm_interface/lib/src/osm_interface.dart
@@ -33,6 +33,8 @@ abstract class OSMPlatform extends PlatformInterface {
 
   Stream<UserLocationEvent> onUserPositionListener(int idMap);
 
+  Stream<RegionIsChangingEvent> onRegionIsChangingListener(int idMap);
+
   Future<void> init(
     int idOSM,
   );
@@ -249,6 +251,10 @@ abstract class MobileOSMPlatform extends OSMPlatform {
   );
 
   Future<void> removeLimitArea(
+    int idOSM,
+  );
+
+  Future<BoundingBox> getBounds(
     int idOSM,
   );
 }